requirements_local.txt
pdfplumber>=0.10.0
openai>=1.30.0
python-dotenv>=1.0.0
tqdm>=4.66.0
huggingface_hub>=0.23.0
